The File
Most probably numerous agencies, companies and individuals who have the capability to break the encryption of this file can hack into it. But that would be impolite and inconsiderate. In this case, an interestd person only needs to ask formally for the key, and then open the file using the given key. Just because one can break-in and enter into a private space without consent, or pilfer a file with ease, doesn't mean that they ought to.