The File
Most probably numerous agencies, companies, and individuals who have the capability to break the encryption of this file can hack into it. Some might have even fished its key by simply spying on my computer activity. But those methods would be impolite and inconsiderate. In this case, an interestd person only needs to ask formally for the key, and then open The File using the given key. Just because one can break-in and enter into a private space without consent, spy, evesdrop, or even pilfer an equitable asset with ease, doesn't mean that they ought to.