justice owed to victims and survivors is long overdue!
some people like to say, "skip the details" or "spare me the details."

so i've only provided the editorial work, and a number of excerpts from the main body of text, in this publicly visibible repository. 

the details along with various proofs, evidence, facts, and data are in the protected files.